Interpreter cleanup
Last round of fixes before I start phasing out the Interpreter for the
p-code based compiler.  There are a bunch of different fixes here, I'll
try and be more organized in the future.

Lambda expressions and macros more or less work, but it looks like the
tail call optimization makes more sense in the context of a compiler.