Define variables
Restore back patch support (via the Undefined type).  If a variable is
undefined where it is called, insert a call to an unspecified
address (which will be patched when the value is defined).

This means defining a constants actually are compiles a thunk which is
certainly not great, but works for now.