Overhaul type conversion
The 'unwrap' function had overloaded semantics, retrieving the 'value'
of immediates and the tag/value version of built-in types.  unwrap has
been re-written to just the reinterpret_cast of the Any input, and the
'value' function has been added, which returns the value field of
immediates.

'unwrap' will still return the pointer to implementation for a Pimple
type.  That may change in the future as well, but I don't find that as
confusing as the overloaded semantics for immediates.