Movable pcode entry point
For a REPL to work I need definitions to persist in the output pcode but
the 'script' part to be over-written or skipped at the top of each loop.
Add a pointer into the pcode so that execution doesn't begin at 'pc=0'
and I can append to/skip over the 'define' section.