Working compiler based REPL
The REPL has a block of PCode which starts with definitions and is
followed by the most recently compiled statement.  A pointer to the most
recent statement is passed to the VM as an entry point.

The statement is evaluated, if the statement was a definition the entry
point is moved to the end of that definition, otherwise it is reset so
the next input statement over-writes the previous one.