Add PrimitiveMacro support
Primitive macros are based on the CxxFunctor type but use the arguments
differently.  The macro will interpret its first argument as a pointer
to an 'Eval' object and the second as the Ast expression (including its
symbol).

The API is currently error prone and a little silly (really wasn't able
to leverage the CxxFunctor common interface).  See issue #3