Add Arena type for fixed lifespan allocation
There are places where I want to allocate a few objects or an Ast for
the duration of a call.  Add an stub Arena class.  The class can
allocate Atl objects, and when it is complete it should free the
allocated objects when the Arena is destroyed.

Pools struct out of GC so they can be used by other allocation models.