Raise error on string with multiple expressions
The ParseString.string_ function returns a single result and does not
mutate its argument so if the string has multiple top-level expressions,
only the first one will ever be parsed.  Add a check to
ParseString.string_ which ignores whitespace and comments but raises an
exception if a string contains more than one sexp.