Cleanup syntax primitives
I'm going to apply some macro transformation to Lambda, so, for now,
it's not going to be a syntax primitive.  Remove the documentation and
special handling from the parser.

Cleanup the special handling for Let as well, and remove the type.  I
may need to add it back if I want polymorphic let type deduction to
work, but that's a ways off from happening.