Use non-macro lambda in test/compile.cpp
I'd like use a macro to do some expansion on lambda during compilation.
The existing compile tests aren't trying to verify macro expansions,
just that the compiler generates useable procedures.  Add __\__ to the
environment mapping to a Lambda and Replace the calls to '\' with calls
to '__\__'.

The compile tests also be as independent of the parser.  Some of them
have been re-written to use the make_ast::make functions.