Explicit pcode termination
After compiling an expr, the compiler does not know if it is waiting for
more expressions, or if the p-code is going to the VM.  Use a new
RunnableCode type to terminates the PCode with a 'finish' instruction.
TinyVM.run accepts RunnableCode to prevent it from getting un-terminated
pcode.

With this in place, macros and won't have to worry about using
'in_pace_any' or pushing/popping finish statements, and the VM.run*
parameter types prevent un-terminated code from being passed in.