Build type declaration into the compiler
I would like to make macros which take an Ast and give a new one back to
the compiler.  Of the "macros" I have, only the type declaration
operator needs to interact more directly with the compiler; put that
logic directly into the compiler rather than make a different interface
for the one operation.