Return Pointer for parsed Ast
The new Ast type stores an offset relative to itself so it cannot be
passed by value meaningfully.  Asts are often cassed to Any and passed
around, but it would be difficult to adapt all the code which uses Any
to pass by reference, so we'll wrap Asts with Pointer and pass that
about.

The parser internals were also adapted to use the AstSubstrate.

Also re-indented the code.  Should have been two commits, but it's too
late now.