Prevent Asts from being passed by value
The wrap/unwrap functions are the principle way Atl data are passed
around.  Complain at compile time if unwrap<Ast> is used on a non-lvalue
references.

Since I can't tell the difference between a pass-by-value and a const
rvalue, unwrap<Ast> will complain in cases where it should technically
work.  In that situation the address of the Any can be taken (ie
unwrap<Ast>(&foo)).