Remove old macro helpers
The original 'primitive-macro' type would be passed a compiler and vm
which could be used to evaluate expressions then discard the compiled
code once the macro had a result.  The helpers for discarding
top-of-code in the compiler and vm are not used anywhere else, so I'm
getting rid of them.