Include a "symbol offset table" in a Code object
The symbol-offset-table keeps a map of symbol names to code positions
and a multi-map of code positions to symbol names.  The latter is aid in
byte-code debugging, the former replaces the 'main_entry_point'
attribute with something more flexible.  It will hopefully also let me
to inter Code code calls.