Replace the RunnableCode type
The RunnableCode type added a couple cleanup instructions to code and
removed them when the code is passed back out of it.  The cleanup
instructions have become a little more complex and depend on whether or
not the code in question is considered a fragment or has a "main"
routine.

Rather than have RunnableCode try to figure what is supposed to be going
on from the byte code, give the Compile and AssembleCode methods for
taking a relinquishing a piece of Code and let them do it.

The Code is passed as a pointer, although it should probably be passed
as an rvalue #35