Assign free variables before compilation
Do an initial pass over the AST to accumulating the list of free vars at
each scope and using that information to generate DefProcedure structs
to replace Lambdas.  I need this because the compiler needs to know how
to build the closure for a function even if the function is lexically
defined after the call.