C++ level pattern matcher
An Ast de-constructor analogous to the make_ast family of functions.
The pattern_matcher functions can be used to produce a "matcher"
function which can confirm that an Ast (or at least its start) conform
to a certain format, and the matcher can extract values from the Ast
while applying.