Add backpatching support to the 'assembler'
To facilitate testing, add support for text labels and replacing an atom
at those labels.

Add support for patching multiple atoms.  The actual implementation
needs to be fleshed out by the compiler; it needs to decide how much
space to leave for the future back-patch and will implement the calls
which generate the over-writing atoms.