Use binary function types for c++ functions
The AlgorithmW implementations I've found use a function constructor is
binary.  I thought it would be cute to build my implementation around a
variadic function constructor (ie `(-> a b c)` instead of `(-> a (-> b
c))`), but I'm not sure it isn't causing problems with the inference.

I want to be able to derive the function signature from a c++ function
spec, and the fn_type::fn helper allows me to do just that.

Update ffi.hpp to use the fn_type::fn helper