utility: Take const tuple in ForEachTuple
Taking its arguments as const makes ForEachTuple in more situations, and
it can mangle the things a tuple points to without modifying the tuple
itself.  Which I find confusing, but there you go.

Tests helped surface a related issue.  I'm sticking them in this commit
because they are tests for the utility file.