type_inference: Apply substitutions to more types
When applying type substitutions we need to update all types referenced
in a scope, not just Symbol schemes.  The Lambda metadata contains a
reference to its return type, and a Type instance (somewhat obviously)
contains a reference to a type.  Both need to have the types they are
wrapping updated.