Build AstData wrapping into Ast interface
The AstData type is special because it can't be passed by data.  It's
also a little annoying because functions which deal with Asts also have
had to deal with AstData types.

Get rid of the pass_value function which wraps AstData if that's what it
gets, and move the wrapping into the Ast traversal helpers.  This
requires some special care if the user is trying to modify the Ast's
data, since modifying the wrapped value won't modify the actual thing
being wrapped.

The ModifyDataIterator can only wrap one AstData at a time, so they
should be ephemeral things.