lexical_environment: functional assign_forms
I'm trying to abstract away the difference between Ast and AstData.  The
abstraction involves wrapping an AstData which means we don't get real
references to all the nodes in an Ast when iterating over it.

There are a couple ways to solve this, but the cleanest seems to be
making a fresh Ast rather than modifying elements of the old one.