Remove LambdaMetadata.pad_to
In general we don't know how much space a tail-called function will
require, so the idea of computing a 'pad_to' value which leave extra
space on the stack is a bit quixotic.  It is also unnecessary because
there is a fixed amount of information on the stack which we need to
save (the caller's return address etc) before over-writing the a callers
arguments.

Store a plane argument count in the run-time closures and get rid of the
'pad_to' attribute in LambdaMetadata.  On the stack, store argument
count around the frame rather than around the return_ instruction (I
don't think I have to, but it seemed a little more convenient).