Define values at runtime
In general define involves assigning the value of an expression to a
symbol.  It's simplest to evaluate the expression at runtime; add an
array to the VM such that each symbol refers to a particular address
in the array.  Using the symbol means getting the value of the array
and vice versa.