helpers: Iterator like Ast::iterator wrapper
The Ast::iterators try to pretend AstData nodes are Ast nodes (that
is, wrapped in a relocatable class).  To do this, their operator*
needs to return a temporary rather than a reference since it must
replace AstData references with Ast temporaries.

This mostly works, but I have some loops which traverse Asts and
modify their atoms.  In that case what I would like is a sequence of
the Ast::iterator instances so I can call the explicit "reference" or
"pointer" methods.  To support that, I added an itritrs wrapper which
makes a range whose de-referenced values are Ast::iterators.