utility: deref or passthough a based on input type
I have some template functions like 'is' which expect a reference
argument.  This is annoying in the context of a template function
which would like to use 'is' but could sensibly take either a pointer
or reference argument type since overloading all my functions (in
particular 'wrap') to work with either pointers or references would be
messy and not always to what I want.

The 'deref' in utility returns a reference to a type by either passing
through a reference argument or dereferencing a pointer argument.
This can be used in a template context to support functions which can
sensibly accept either argument type and would like to use 'is',
'wrap', et al.