type: Use Ast as closure parameter container
The LambdaMetadata type needs to keep track of where it's free
variables are coming from and pass that information to the compiler so
the compiler can setup the closure information into a function call.
LambdaMetadata doesn't know about free variables up front, so the
data-structure holding them must be append-able.  At the same time,
the compiler deals with Ast::iterators to side-step memory management
issues during garbage collection.  So; we collect free-variable
information into a std::vector and use that to build an Ast which the
compiler can use later in the compilation process.