lexical_environment: Find values before assigning forms
The 'assign_form' function copys an Ast processing some special forms
and generating function metadata.  For this to work properly it needs
to identify symbols which have the value of a special form (ie, map
"define" to the Define type).

Add a WalkValues AST walker which substitutes some values in for
symbols and use it in the AssignForms class.