utility: Cleanup BuildIndicies
The BuildIndicies type creates an 'Indexer' which can index an
argument pack of a given size.  It had been inheriting 'Indexer',
which sometimes works, but sometimes breaks for partially specialized
templates which expect an Indexer class.

Nest the Indexer as the BuildIndicies::type and explicitly access the
nested class to avoid ambiguity.