print: Distinguish between Ast and AstData when printing with_type
There can be Ast's embedded in AstData; they are supposed to be
abstracted by the Ast::iterator/AstWalker types, so printing has
displayed them uniformly.  Because the abstractions don't always
apply (or work correctly) use different brace styles to indicate which
is which when printing with_type to facilitate debugging.